# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Holy Hymns API",
    "version": "1.0.0",
    "description": "Self-hosted lyrics catalogue. Session authorization is checked on every protected request. All timestamps UTC. Public responses never contain drafts."
  },
  "servers": [
    {
      "url": "http://localhost:8080/v1",
      "description": "Local development"
    }
  ],
  "paths": {
...
    "/auth/register": {
      "post": {
        "summary": "Register",
        "security": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or revoked session"
          },
          "403": {
            "description": "Role or origin denied"
          },
          "404": {
            "description": "Resource unavailable"
          },
          "409": {
            "description": "Version or identity conflict"
          },
          "429": {
            "description": "Rate or free email quota exhausted"
          },
          "503": {
            "description": "Dependency unavailable"
          },
          "500": {
            "description": "Unexpected server or database failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "maxLength": 254
                  },
                  "password": {
                    "type": "string",
                    "minLength": 12,
                    "description": "At least 12 Unicode characters and at most 1024 UTF-8 bytes."
                  },
                  "name": {
                    "type": "string",
                    "description": "Optional display name; at most 200 UTF-8 bytes."
                  }
                },
                "required": [
                  "email",
                  "password"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "description": "Requires configured SMTP. Registration queues email verification; it does not return a session. A repeated registration never replaces existing credentials. Omit name to use an empty display name."
      }
    },
    "/auth/verify": {
      "post": {
        "summary": "Verify",
        "security": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or revoked session"
          },
          "403": {
            "description": "Role or origin denied"
          },
          "404": {
            "description": "Resource unavailable"
          },
          "409": {
            "description": "Version or identity conflict"
          },
          "429": {
            "description": "Rate or free email quota exhausted"
          },
          "503": {
            "description": "Dependency unavailable"
          },
          "500": {
            "description": "Unexpected server or database failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string"
                  }
                },
                "required": [
                  "token"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "description": "Consume the single-use email-verification token; it expires after 24 hours. Sign in separately after success."
      }
    },
    "/auth/forgot-password": {
      "post": {
        "summary": "Forgot password",
        "security": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or revoked session"
          },
          "403": {
            "description": "Role or origin denied"
          },
          "404": {
            "description": "Resource unavailable"
          },
          "409": {
            "description": "Version or identity conflict"
          },
          "429": {
            "description": "Rate or free email quota exhausted"
          },
          "503": {
            "description": "Dependency unavailable"
          },
          "500": {
            "description": "Unexpected server or database failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  }
                },
                "required": [
                  "email"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "description": "Requires configured SMTP. Returns the same generic message for missing and eligible accounts. Reset links expire after one hour."
      }
    },
    "/auth/reset-password": {
      "post": {
        "summary": "Reset password",
        "security": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or revoked session"
          },
          "403": {
            "description": "Role or origin denied"
          },
          "404": {
            "description": "Resource unavailable"
          },
          "409": {
            "description": "Version or identity conflict"
          },
          "429": {
            "description": "Rate or free email quota exhausted"
          },
          "503": {
            "description": "Dependency unavailable"
          },
          "500": {
            "description": "Unexpected server or database failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 12,
                    "description": "At least 12 Unicode characters and at most 1024 UTF-8 bytes."
                  }
                },
                "required": [
                  "token",
                  "password"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "description": "Consume a one-hour single-use reset token, replace the password, and revoke existing sessions. Sign in separately afterward."
      }
    },
    "/auth/accept-invitation": {
      "post": {
        "summary": "Accept invitation",
        "security": [
          {},
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or revoked session"
          },
          "403": {
            "description": "Role or origin denied"
          },
          "404": {
            "description": "Resource unavailable"
          },
          "409": {
            "description": "Version or identity conflict"
          },
          "429": {
            "description": "Rate or free email quota exhausted"
          },
          "503": {
            "description": "Dependency unavailable"
          },
          "500": {
            "description": "Unexpected server or database failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string",
                    "description": "Required only for a new account; minimum 12 Unicode characters, maximum 1024 UTF-8 bytes."
                  },
                  "name": {
                    "type": "string",
                    "description": "Optional for a new account, at most 200 UTF-8 bytes. Existing account name is preserved."
                  }
                },
                "required": [
                  "token"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "description": "Invitation tokens expire after 48 hours. Existing accounts must authenticate with the invited account; their password/name are preserved. New accounts supply a password. Success grants admin access and revokes existing sessions; sign in afterward. This endpoint returns a message, not a session."
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Sign in with verified email and password",
        "security": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or revoked session"
          },
          "403": {
            "description": "Role or origin denied"
          },
          "404": {
            "description": "Resource unavailable"
          },
          "409": {
            "description": "Version or identity conflict"
          },
          "429": {
            "description": "Rate or free email quota exhausted"
          },
          "503": {
            "description": "Dependency unavailable"
          },
          "500": {
            "description": "Unexpected server or database failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "email",
                  "password"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "description": "Only verified, active accounts may sign in. Response token is an opaque revocable session. Attempts are limited by trusted client IP and account."
      }
    },
    "/auth/me": {
      "get": {
        "summary": "Read active current account",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or revoked session"
          },
          "403": {
            "description": "Role or origin denied"
          },
          "404": {
            "description": "Resource unavailable"
          },
          "409": {
            "description": "Version or identity conflict"
          },
          "429": {
            "description": "Rate or free email quota exhausted"
          },
          "503": {
            "description": "Dependency unavailable"
          },
          "500": {
            "description": "Unexpected server or database failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Revoke current session",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or revoked session"
          },
          "403": {
            "description": "Role or origin denied"
          },
          "404": {
            "description": "Resource unavailable"
          },
          "409": {
            "description": "Version or identity conflict"
          },
          "429": {
            "description": "Rate or free email quota exhausted"
          },
          "503": {
            "description": "Dependency unavailable"
          },
          "500": {
            "description": "Unexpected server or database failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout-all": {
      "post": {
        "summary": "Revoke all sessions",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or revoked session"
          },
          "403": {
            "description": "Role or origin denied"
          },
          "404": {
            "description": "Resource unavailable"
          },
          "409": {
            "description": "Version or identity conflict"
          },
          "429": {
            "description": "Rate or free email quota exhausted"
          },
          "503": {
            "description": "Dependency unavailable"
          },
          "500": {
            "description": "Unexpected server or database failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/auth/account": {
      "delete": {
        "summary": "Delete account after recent reauthentication; owner protected",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "manualRevocationUrl": {
                      "type": "string",
                      "format": "uri",
                      "description": "Present when an older linked Apple account has no stored grant. Local deletion succeeded; the user should remove the remaining Apple authorization using this URL."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or revoked session"
          },
          "403": {
            "description": "Role or origin denied"
          },
          "404": {
            "description": "Resource unavailable"
          },
          "409": {
            "description": "Version or identity conflict"
          },
          "429": {
            "description": "Rate or free email quota exhausted"
          },
          "503": {
            "description": "Dependency unavailable"
          },
          "500": {
            "description": "Unexpected server or database failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "password": {
                    "type": "string",
                    "description": "Required when the account has a password; otherwise omit and send {} after a fresh social sign-in."
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "description": "Requires a JSON body. Password-bearing accounts must supply the current password. Social-only accounts must have signed in within the last ten minutes and may send {}. Provider, idToken and nonce are not accepted here. The last active verified owner cannot delete their account. Success deletes personal identity/session/favorite data; anonymized audit records and aggregate analytics remain. If Apple is linked and a grant is stored, the server revokes it before deleting local data; revocation failure leaves the account intact for retry. Historical Apple accounts with no stored grant are deleted and return manualRevocationUrl so the user can remove the remaining Apple authorization."
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Opaque revocable account session, not a provider identity token."
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "reader",
              "admin",
              "owner"
            ]
          },
          "verified": {
            "type": "boolean"
          },
          "suspended": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "role",
          "verified",
          "suspended"
        ]
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
      "Message": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
      "Session": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "required": [
          "token",
          "user"
        ]
      },
...
    }
  }
}